test(metadata): ♻️ переиспользовать регистрацию явного XML
